notebook/unstr.py
from langchain_google_genai import ChatGoogleGenerativeAI
from dotenv import load_dotenv
import os

load_dotenv()

llm = ChatGoogleGenerativeAI(model="gemini-1.5-flash", temperature=0.3)
response = llm.invoke("Say Hello, Gemini how are you!")
print(response.content)