requirements.txt
langchain_astradb
pandas
langchain_google_genai
uvicorn 
python-multipart
jinja2
python-dotenv 
langchain
langchain_core
langchain_pinecone
langgraph
tavily-python
langchain_community
streamlit
fastapi[all]
langchain-pinecone
pypdf
streamlit
openai
langchain_groq
-U langchain-community

-e .